requirements.txt
fastapi
uvicorn
pymongo
requests
python-dotenv
emoji
PyJWT